Explicitly implement foldMap' for V1 through V4
The impetus for doing so is the doctests for `Trace`, which invokes:

```hs
>>> trace (V2 (V2 a b) (V2 c d))
a + d
```

Where `trace` is implemented as `sum . diagonal`. In GHC 9.0 and earlier,
`sum` is implemented in terms of `foldMap`, whereas in 9.2 and later, it is
implemented in terms of `foldMap'`. Because `foldMap` has an explicit
definition for `V2`:

```hs
foldMap f (V2 a b) = f a `mappend` f b
```

`mempty` is never used, so the `simple-reflect` expression above will evaluate
to `a + d`. On the other hand, `foldMap'` did not have an explicit
implementation, so it relied on its default implementation in terms of
`foldl'`:

```hs
foldMap' f = foldl' (\ acc a -> acc <> f a) mempty
```

The use of `mempty` there would cause the doctests to print `0 + a + d` instead
of `a + d`. It's a rather minor thing in the grand scheme of things, but I'd
prefer the doctests to print out the same thing on all versions of GHC that we
test. Explicitly implementing `foldMap'` is the easiest way to guarantee that,
and the resulting code is probably slightly easier to optimize as well.

We could conceivably implement `foldMap'` explicitly for all `Foldable`
instances in `linear`, but I will leave that as future work.